testcases: add cli-args for runtime instrumentation
Adjust to removal of URI-encoding

The URI-encoding provided runtime-instrumentation via the
device-uri/path. This has been removed in favor of providing actual
arguments in the command-line-interface and via 'xnvme_opts' in the API.

This commit adjust to these changes by providing cli-arguments in
testcases, updating the reference environment, and adjusting the
testplans.

* envs: remove URI-encoding
* testcases: add cli-args for runtime instrumentation
* testsuites: remove 'xnvme_tests_ident"

Removed "zoned_mgmt.sh" as "xnvme_tests_znd_state.sh" replaces it.

Signed-off-by: Simon A. F. Lund <simon.lund@samsung.com>